Fixing and making it work
Only updating progress left to get working!

+++ src/OpenHomeServer.Server/Plugins/Ripper/RipperNotificationHub.cs
@@ -17,7 +17,7 @@ namespace OpenHomeServer.Server.Plugins.Ripper
             _hubContext = hubContextFactory.CreateHubContext<RipperNotificationHub>();
         }
 
-        public void UpdateProgress(int trackNumber, int percentageComplete)
+        public void UpdateProgress(int trackNumber, double percentageComplete)
         {
             _hubContext.Clients.All.onRippingProgress(trackNumber, percentageComplete);
         }
